local_script_test/requirements.txt
beautifulsoup4
requests
langchain
langchain-community
sentence-transformers
weaviate-client
torch
langchain-text-splitters
minio